pa1/crawled_data_inspection/requirements.txt
PyYAML
numpy
pandas
tqdm
stanza
scikit-learn
umap-learn
keybert
plotly
psycopg2-binary
pyyaml 
beautifulsoup4 
lxml